--- blueocean-commons/src/test/java/io/jenkins/blueocean/commons/stapler/ModelObjectSerializerTest.java
@@ -1,55 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright (c) 2016, CloudBees, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package io.jenkins.blueocean.commons.stapler;
-
-import net.sf.json.JSONObject;
-import org.junit.Assert;
-import org.junit.Test;
-import org.kohsuke.stapler.export.Exported;
-import org.kohsuke.stapler.export.ExportedBean;
-
-import java.io.IOException;
-
-/**
- * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
- */
-public class ModelObjectSerializerTest {
-
-    @Test
-    public void test_json() throws IOException {
-        String xJson = ModelObjectSerializer.toJson(new X());
-        JSONObject jsonObj = JSONObject.fromObject(xJson);
-
-        Assert.assertEquals(X.class.getName(), jsonObj.getString("_class"));
-        Assert.assertEquals("xVal", jsonObj.getString("val"));
-    }
-
-    @ExportedBean
-    public static class X {
-        @Exported
-        public String val() {
-            return "xVal";
-        }
-    }
-}
