+++ blueocean-dashboard/src/main/java/io/jenkins/blueocean/preload/PipelineActivityStatePreloader.java
@@ -27,7 +27,7 @@ import hudson.Extension;
 import hudson.model.Item;
 import io.jenkins.blueocean.commons.BlueUrlTokenizer;
 import io.jenkins.blueocean.commons.RESTFetchPreloader;
-import io.jenkins.blueocean.commons.stapler.ModelObjectSerializer;
+import io.jenkins.blueocean.commons.stapler.Export;
 import io.jenkins.blueocean.rest.model.BluePipeline;
 import io.jenkins.blueocean.rest.model.Container;
 import io.jenkins.blueocean.rest.model.Resource;
@@ -69,7 +69,7 @@ public class PipelineActivityStatePreloader extends RESTFetchPreloader {
             while(activitiesIterator.hasNext()) {
                 Resource blueActivity = activitiesIterator.next();
                 try {
-                    activities.add(JSONObject.fromObject(ModelObjectSerializer.toJson(blueActivity)));
+                    activities.add(JSONObject.fromObject(Export.toJson(blueActivity)));
                 } catch (IOException e) {
                     LOGGER.log(Level.FINE, String.format("Unable to preload runs for Job '%s'. Activity serialization error.", pipeline.getFullName()), e);
                     return null;
